Use subcourse id instead of course id to create bbb meetings and get bbb meeting urls. Fix wrong apidoc (see issue #113).
