+++ web/index.ts
@@ -136,8 +136,8 @@ createConnection().then(() => {
         coursesRouter.put("/:id/subcourse/:subid/lecture/:lecid", courseController.putLectureHandler);
         coursesRouter.delete("/:id/subcourse/:subid/lecture/:lecid", courseController.deleteLectureHandler);
 
-        coursesRouter.get("/:id/meeting", courseController.getCourseMeetingHandler);
-        coursesRouter.get("/:id/meeting/join", courseController.joinCourseMeetingHandler);
+        coursesRouter.get("/:id/subcourse/:subid/meeting", courseController.getCourseMeetingHandler);
+        coursesRouter.get("/:id/subcourse/:subid/meeting/join", courseController.joinCourseMeetingHandler);
 
         app.use("/api/course", coursesRouter);
     }
